Add 'format' subcmd, add pickle output to 'report'
The '--format pickle' option to 'report' subcommand makes the output
to be the raw data collected from the APIs (called a modedict in the
code) in Python pickle format. The output file can be set with the
'--output'/'-o' option.

The format subcommand outputs a formatted report (or another pickle
file) when given a pickle output from 'taival report' as input.